Remove coordinate_system.py (#18)
